Add a --select-from-relations option.
This runs a SELECT * FROM <table> LIMIT 0 query on each table,
TOAST table, and materialized view in the database, which gives us
some chance of finding out whether parts of the database are
completely inaccessible. It doesn't help with indexes, or relation
segments other than the first one, but it's something.

Rushabh Lathia and Robert Haas