Avoid running out of memory in the face of large catalog tables.
If a table has no dependencies (e.g. foreign keys that point to it),
then we need not store the catalog data in an in-memory hash table,
and we can read and check rows one at a time using single-row mode
instead of reading the result set all at once. Thus turns out to
be faster, because the old coding built a bunch of hash tables that
we didn't really need, and it also saves a lot of memory.

Typically, we don't expect catalog tables to be large enough for
this to matter, but it could, especially for pg_largeobject.

Patch by Shruthi KC, reviewed by Mahendra Singh Thalor.